try vendoring SFML to figure out what's wrong
just kidding, they have a massive window implementation for Unix/X11
and it's kind of messy. All of it is reliant on X11 so the only way
I'd be able to interface with it is I redid all of the work in
GLFW. It would probably work alright, but it would take a lot longer
than just literally making my own OpenGL wrapper.